pyproject.toml
[tool.poetry]
name = "aether"
version = "0.0.1"
description = "An experimental, fast & robust library for Discord."
authors = ["18 <111544899+eightween@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
aenum = "^3.1.15"
trio = "^0.24.0"
trio-websocket = "^0.11.1"
attrs = "^23.2.0"
cattrs = "^23.2.3"
ujson = "^5.9.0"
python-dotenv = "^1.0.1"
tabulate = "^0.9.0"

[tool.poetry.dev-dependencies]
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
pre-commit = "^3.6.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
